requirements.txt
streamlit
groq
python-dotenv
matplotlib
simplejson
PyMuPDF
